requirements.txt
Flask==3.0.3
Flask-Restx==1.3.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
PyJWT==2.9.0
Marshmallow==3.22.0

Pytest==8.3.3
Pytest-Flask==1.3.0
